+++ lib/grid_codec/types/char_array.ex
@@ -17,11 +17,13 @@ defmodule GridCodec.Types.CharArray do
   Then use it in your codec:
 
       defmodule MyApp.OrderEvent do
-        use GridCodec.Struct, types: [symbol: MyApp.Types.Symbol]
+        use GridCodec.Struct
+
+        alias MyApp.Types.Symbol
 
         defcodec do
           field :order_id, :u64
-          field :symbol, :symbol
+          field :symbol, Symbol
         end
       end